backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
pyjwt==2.8.0
bcrypt==4.1.2
scikit-learn>=1.3.2
numpy==1.26.2
